pyproject.toml
[tool.poetry]
name = "upworktasktests"
version = "0.1.0"
description = ""
authors = ["Ivan <korneichuk.ik@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
webdriver-manager = "^4.0.0"
bs4 = "^0.0.1"
requests-html = "^0.10.0"
pandas = "^2.0.3"
html5lib = "^1.1"
moviepy = "^1.0.3"
matplotlib = "^3.7.2"
selenium = "4.8.0"
selenium-recaptcha-solver = "^1.9.0"
pytesseract = "^0.3.10"
pillow = "^10.0.0"
opencv-python = "^4.8.0.76"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
